requirements.txt
python-nmap
python-libnmap
netaddr
configparser
python-crontab